feat: setting Page 컴포넌트 분리
